+++ vips/arithmetic.c
@@ -1,0 +1,17 @@
+#include "arithmetic.h"
+
+int add(VipsImage *left, VipsImage *right, VipsImage **out) {
+	return vips_add(left, right, out, NULL);
+}
+
+int multiply(VipsImage *left, VipsImage *right, VipsImage **out) {
+	return vips_multiply(left, right, out, NULL);
+}
+
+int linear1(VipsImage *in, VipsImage **out, double a, double b) {
+	return vips_linear1(in, out, a, b, NULL);
+}
+
+int invert_image(VipsImage *in, VipsImage **out) {
+	return vips_invert(in, out, NULL);
+}
