+++ vips/bridge.go
@@ -2,20 +2,11 @@ package vips
 
 // #cgo pkg-config: vips
 // #include "bridge.h"
+// #include "foreign.h"
 import "C"
 import (
 	"fmt"
-	"log"
-	"math"
-	"runtime"
-	dbg "runtime/debug"
 	"unsafe"
-)
-
-const (
-	defaultQuality     = 90
-	defaultCompression = 6
-	maxScaleFactor     = 10
 )
 
 type vipsLabelOptions struct {
@@ -30,10 +21,6 @@ type vipsLabelOptions struct {
 	Margin    C.int
 	Opacity   C.float
 	Color     [3]C.double
-}
-
-type vipsLoadOptions struct {
-	cOpts C.ImageLoadOptions
 }
 
 func vipsCall(name string, options []*Option) error {
@@ -85,54 +72,6 @@ func vipsCallOperation(operation *C.VipsOperation, options []*Option) error {
 	return nil
 }
 
-func vipsHasProfile(in *C.VipsImage) bool {
-	return int(C.has_profile_embed(in)) > 0
-}
-
-func vipsLoadFromBuffer(buf []byte, opts ...LoadOption) (*C.VipsImage, ImageType, error) {
-	// Reference buf here so it's not garbage collected during image initialization.
-	defer runtime.KeepAlive(buf)
-
-	var out *C.VipsImage
-
-	imageType := vipsDetermineImageType(buf)
-	if imageType == ImageTypeUnknown {
-		if len(buf) > 2 {
-			log.Printf("Failed to understand image format size=%d %x %x %x", len(buf), buf[0], buf[1], buf[2])
-		} else {
-			log.Printf("Failed to understand image format size=%d", len(buf))
-		}
-		return nil, ImageTypeUnknown, ErrUnsupportedImageFormat
-	}
-
-	bufLength := C.size_t(len(buf))
-	imageBuf := unsafe.Pointer(&buf[0])
-
-	var loadOpts vipsLoadOptions
-	for _, opt := range opts {
-		opt(&loadOpts)
-	}
-
-	err := C.init_image(imageBuf, bufLength, C.int(imageType), &loadOpts.cOpts, &out)
-	if err != 0 {
-		return nil, ImageTypeUnknown, handleVipsError(out)
-	}
-
-	return out, imageType, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-conversion.html#vips-copy
-func vipsCopyImage(in *C.VipsImage) (*C.VipsImage, error) {
-	var out *C.VipsImage
-
-	err := C.copy_image(in, &out)
-	if int(err) != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
 func vipsExportBuffer(image *C.VipsImage, params *ExportParams) ([]byte, ImageType, error) {
 	tmpImage, err := vipsPrepareForExport(image, params)
 	if err != nil {
@@ -157,8 +96,8 @@ func vipsExportBuffer(image *C.VipsImage, params *ExportParams) ([]byte, ImageTy
 		return nil, ImageTypeUnknown, fmt.Errorf("cannot save to %#v", ImageTypes[format])
 	}
 
-	if params.BackgroundColor != nil {
-		tmpImage, err = vipsFlattenBackground(tmpImage, *params.BackgroundColor)
+	if params.BackgroundColor != nil && vipsHasAlpha(tmpImage) {
+		tmpImage, err = vipsFlatten(tmpImage, params.BackgroundColor)
 		if err != nil {
 			return nil, ImageTypeUnknown, err
 		}
@@ -265,239 +204,6 @@ func vipsDetermineImageType(buf []byte) ImageType {
 	return ImageTypeUnknown
 }
 
-func vipsFlattenBackground(in *C.VipsImage, color Color) (*C.VipsImage, error) {
-	incOpCounter("flatten")
-	var out *C.VipsImage
-
-	if int(C.has_alpha_channel(in)) > 0 {
-
-		bg := [3]C.double{
-			C.double(color.R),
-			C.double(color.G),
-			C.double(color.B),
-		}
-
-		err := C.flatten_image_background(in, &out, bg[0], bg[1], bg[2])
-		if int(err) != 0 {
-			return nil, handleVipsError(out)
-		}
-		unrefImage(in)
-
-		in = out
-	}
-
-	return in, nil
-}
-
-// Resize executes the 'resize' operation
-func vipsResize(in *C.VipsImage, scale, vscale float64, kernel Kernel) (*C.VipsImage, error) {
-	incOpCounter("resize")
-	var out *C.VipsImage
-
-	// Let's not be insane
-	scale = math.Min(scale, maxScaleFactor)
-	vscale = math.Min(vscale, maxScaleFactor)
-
-	if err := C.resize_image(in, &out, C.double(scale), C.double(vscale), C.int(kernel)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsEmbed(in *C.VipsImage, left, top, width, height int, extend ExtendStrategy) (*C.VipsImage, error) {
-	incOpCounter("embed")
-	var out *C.VipsImage
-
-	if err := C.embed_image(in, &out, C.int(left), C.int(top), C.int(width), C.int(height), C.int(extend), 0, 0, 0); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-conversion.html#vips-autorot
-func vipsAutoRotate(in *C.VipsImage) (*C.VipsImage, error) {
-	incOpCounter("autorot")
-	var out *C.VipsImage
-
-	if err := C.autorot_image(in, &out); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-conversion.html#vips-autorot
-func vipsRotate(in *C.VipsImage, angle Angle) (*C.VipsImage, error) {
-	incOpCounter("rot")
-	var out *C.VipsImage
-
-	if err := C.rot_image(in, &out, C.VipsAngle(angle)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsZoom(in *C.VipsImage, xFactor, yFactor int) (*C.VipsImage, error) {
-	incOpCounter("zoom")
-	var out *C.VipsImage
-
-	if err := C.zoom_image(in, &out, C.int(xFactor), C.int(yFactor)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-conversion.html#vips-composite
-func vipsComposite(ins []*C.VipsImage, mode BlendMode) (*C.VipsImage, error) {
-	incOpCounter("composite")
-	var out *C.VipsImage
-
-	if err := C.composite(&ins[0], &out, C.int(len(ins)), C.int(mode)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsBandJoin(ins []*C.VipsImage) (*C.VipsImage, error) {
-	incOpCounter("bandJoin")
-	var out *C.VipsImage
-
-	if err := C.bandjoin(&ins[0], &out, C.int(len(ins))); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsHasAlpha(in *C.VipsImage) bool {
-	return int(C.has_alpha_channel(in)) > 0
-}
-
-func vipsAddAlpha(in *C.VipsImage) (*C.VipsImage, error) {
-	incOpCounter("addAlpha")
-	var out *C.VipsImage
-
-	if err := C.add_alpha(in, &out); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsAdd(left *C.VipsImage, right *C.VipsImage) (*C.VipsImage, error) {
-	incOpCounter("add")
-	var out *C.VipsImage
-
-	defer unrefImage(left)
-	defer unrefImage(right)
-
-	if err := C.add(left, right, &out); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsMultiply(left *C.VipsImage, right *C.VipsImage) (*C.VipsImage, error) {
-	incOpCounter("multiply")
-	var out *C.VipsImage
-
-	defer unrefImage(left)
-	defer unrefImage(right)
-
-	if err := C.multiply(left, right, &out); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-conversion.html#vips-extract-band
-func vipsExtractBand(in *C.VipsImage, band, num int) (*C.VipsImage, error) {
-	incOpCounter("extractBand")
-	var out *C.VipsImage
-
-	if err := C.extract_band(in, &out, C.int(band), C.int(num)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-//  https://libvips.github.io/libvips/API/current/libvips-arithmetic.html#vips-linear1
-func vipsLinear1(in *C.VipsImage, a, b float64) (*C.VipsImage, error) {
-	incOpCounter("linear1")
-	var out *C.VipsImage
-
-	if err := C.linear1(in, &out, C.double(a), C.double(b)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsExtractArea(in *C.VipsImage, left, top, width, height int) (*C.VipsImage, error) {
-	incOpCounter("extractArea")
-	var out *C.VipsImage
-
-	if err := C.extract_image_area(in, &out, C.int(left), C.int(top), C.int(width), C.int(height)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsFlip(in *C.VipsImage, dir Direction) (*C.VipsImage, error) {
-	incOpCounter("flip")
-	var out *C.VipsImage
-
-	if err := C.flip_image(in, &out, C.int(dir)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-func vipsInvert(in *C.VipsImage) (*C.VipsImage, error) {
-	incOpCounter("invert")
-	var out *C.VipsImage
-
-	if err := C.invert_image(in, &out); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-convolution.html#vips-gaussblur
-func vipsGaussianBlur(in *C.VipsImage, sigma float64) (*C.VipsImage, error) {
-	incOpCounter("gaussblur")
-	var out *C.VipsImage
-
-	if err := C.gaussian_blur(in, &out, C.double(sigma)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
-// https://libvips.github.io/libvips/API/current/libvips-convolution.html#vips-sharpen
-func vipsSharpen(in *C.VipsImage, sigma float64, x1 float64, m2 float64) (*C.VipsImage, error) {
-	incOpCounter("sharpen")
-	var out *C.VipsImage
-
-	if err := C.sharpen(in, &out, C.double(sigma), C.double(x1), C.double(m2)); err != 0 {
-		return nil, handleVipsError(out)
-	}
-
-	return out, nil
-}
-
 func vipsLabel(in *C.VipsImage, params *LabelParams) (*C.VipsImage, error) {
 	incOpCounter("label")
 	var out *C.VipsImage
@@ -532,17 +238,4 @@ func vipsLabel(in *C.VipsImage, params *LabelParams) (*C.VipsImage, error) {
 	}
 
 	return out, nil
-}
-
-func handleVipsError(out *C.VipsImage) error {
-	if out != nil {
-		unrefImage(out)
-	}
-
-	s := C.GoString(C.vips_error_buffer())
-	C.vips_error_clear()
-
-	stack := string(dbg.Stack())
-
-	return fmt.Errorf("%v\nStack:\n%s", s, stack)
 }
