+++ vips/color.go
@@ -1,0 +1,3 @@
+package vips
+
+// #cgo pkg-config: vips
