+++ vips/color.h
@@ -1,0 +1,7 @@
+// https://libvips.github.io/libvips/API/current/libvips-colour.html
+
+#include <stdlib.h>
+#include <vips/vips.h>
+
+int is_colorspace_supported(VipsImage *in);
+int to_colorspace(VipsImage *in, VipsImage **out, VipsInterpretation space);
