+++ vips/convolution.go
@@ -1,0 +1,29 @@
+package vips
+
+// #cgo pkg-config: vips
+// #include "convolution.h"
+import "C"
+
+// https://libvips.github.io/libvips/API/current/libvips-convolution.html#vips-gaussblur
+func vipsGaussianBlur(in *C.VipsImage, sigma float64) (*C.VipsImage, error) {
+	incOpCounter("gaussblur")
+	var out *C.VipsImage
+
+	if err := C.gaussian_blur_image(in, &out, C.double(sigma)); err != 0 {
+		return nil, handleVipsError(out)
+	}
+
+	return out, nil
+}
+
+// https://libvips.github.io/libvips/API/current/libvips-convolution.html#vips-sharpen
+func vipsSharpen(in *C.VipsImage, sigma float64, x1 float64, m2 float64) (*C.VipsImage, error) {
+	incOpCounter("sharpen")
+	var out *C.VipsImage
+
+	if err := C.sharpen_image(in, &out, C.double(sigma), C.double(x1), C.double(m2)); err != 0 {
+		return nil, handleVipsError(out)
+	}
+
+	return out, nil
+}
