+++ vips/header.h
@@ -1,0 +1,7 @@
+// https://libvips.github.io/libvips/API/current/libvips-header.html
+
+#include <stdlib.h>
+#include <vips/vips.h>
+
+
+unsigned long has_profile_embed(VipsImage *in);
